Create custom error class for "Not Found" BZs
Previously, when a BZ was attempted to be cloned, and the origin BZ was
not found, the code continues on as if one was found until it receives a
`NoMethodError` because the code tries calling `[:id]` on `nil`.

This catches that scenario and raises a much more user friendly error,
and updates the specs to reflect the change.